requirements.local.txt
pylint==2.15.5
pylint-pydantic==0.1.4
autopep8==2.0.0
